+++ addOns/dev/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Auth page where the return key does not submit the form
 - Auth page which uses one request and one cookie
 - Auth page which uses multiple requests and multiple cookies
+- OpenAPI auth and unauth pages
 
 ### Changed
 - Update minimum ZAP version to 2.13.0.